market-history-service 추가중 - influxDB 어렵네

+++ mansour/services/market-data-service/src/main/kotlin/org/mj/mansour/marketdata/kis/dto/KisRealtimePriceData.kt
@@ -57,7 +57,7 @@ data class KisRealtimePriceData(
 
     /** 체결 거래량 */
     @param:JsonProperty("CNTG_VOL")
-    val tradeVolume: BigDecimal,
+    val tradeVolume: Long,
 
     /** 누적 거래량 */
     @param:JsonProperty("ACML_VOL")
